test(ui): align sidebar review coverage
