+++ ui/src/e2e/native-nav-responsive-layout.e2e.test.ts
@@ -1,0 +1,155 @@
+import type { BrowserContext } from "playwright";
+import { afterEach, expect, it } from "vitest";
+import { installMockGateway } from "../test-helpers/control-ui-e2e.ts";
+import { createControlUiE2eSuite } from "./control-ui-e2e-suite.test-support.ts";
+import { installNativeWebChrome } from "./native-nav.test-support.ts";
+
+const suite = createControlUiE2eSuite({
+  name: "Control UI native-nav responsive layout E2E",
+  startServerBeforeBrowser: true,
+  unavailableMessage: (executablePath) => `Playwright Chromium is unavailable at ${executablePath}`,
+});
+
+let context: BrowserContext | undefined;
+
+suite.define(() => {
+  afterEach(async () => {
+    await context?.close();
+    context = undefined;
+  });
+
+  async function openPage(options: {
+    hasTouch?: boolean;
+    height?: number;
+    webChrome?: boolean;
+    width?: number;
+  }) {
+    context = await suite.browser.newContext({
+      hasTouch: options.hasTouch,
+      locale: "en-US",
+      serviceWorkers: "block",
+      viewport: { height: options.height ?? 900, width: options.width ?? 1280 },
+    });
+    const page = await context.newPage();
+    if (options.webChrome) {
+      await installNativeWebChrome(page);
+    }
+    await installMockGateway(page, {
+      featureMethods: ["chat.metadata", "chat.startup", "sessions.create"],
+    });
+    const response = await page.goto(suite.server.baseUrl);
+    expect(response?.status()).toBe(200);
+    await page.locator(".sidebar-brand").waitFor({ state: "attached" });
+    return page;
+  }
+
+  it("keeps only history controls in the Settings titlebar", async () => {
+    const page = await openPage({ webChrome: true });
+    const response = await page.goto(`${suite.server.baseUrl}settings/general`);
+    expect(response?.status()).toBe(200);
+
+    const toolbar = page.locator(".macos-titlebar-controls");
+    await expect.poll(() => toolbar.isVisible()).toBe(true);
+    await expect.poll(() => toolbar.getByRole("button").count()).toBe(2);
+    await expect.poll(() => toolbar.getByRole("button", { name: "Back" }).isVisible()).toBe(true);
+    await expect
+      .poll(() => toolbar.getByRole("button", { name: "Forward" }).isVisible())
+      .toBe(true);
+    await expect
+      .poll(() => toolbar.getByRole("button", { name: "Expand sidebar" }).count())
+      .toBe(0);
+    await expect
+      .poll(() => toolbar.getByRole("button", { name: "Open command palette" }).count())
+      .toBe(0);
+    await expect.poll(() => toolbar.getByRole("button", { name: "New session" }).count()).toBe(0);
+  });
+
+  it("keeps the document root scroll-locked in the Settings takeover", async () => {
+    const page = await openPage({ webChrome: true });
+    const response = await page.goto(`${suite.server.baseUrl}settings/general`);
+    expect(response?.status()).toBe(200);
+    await page.locator(".settings-sidebar").waitFor({ state: "visible" });
+
+    // WKWebView scrolls the document whenever it overflows, dragging the
+    // settings sidebar and content along. Force overflow the way stray
+    // content would, then confirm the root refuses to move.
+    const metrics = await page.evaluate(() => {
+      const spacer = document.createElement("div");
+      spacer.style.height = "3000px";
+      document.body.append(spacer);
+      window.scrollTo(0, 500);
+      document.documentElement.scrollTop = 500;
+      document.body.scrollTop = 500;
+      return {
+        bodyScrollTop: document.body.scrollTop,
+        htmlScrollTop: document.documentElement.scrollTop,
+        rootScrollY: window.scrollY,
+      };
+    });
+    expect(metrics).toEqual({ bodyScrollTop: 0, htmlScrollTop: 0, rootScrollY: 0 });
+  });
+
+  it("keeps drawer and search reachable from the narrow chat title bar", async () => {
+    const page = await openPage({ width: 900 });
+    const header = page.locator(".chat-pane__header").first();
+    await expect
+      .poll(() => page.locator(".shell").getAttribute("class"))
+      .toContain("shell--merged-chat-chrome");
+    await expect.poll(() => page.locator(".topbar").isVisible()).toBe(false);
+    await expect
+      .poll(() => header.getByRole("button", { name: "Expand sidebar" }).isVisible())
+      .toBe(true);
+    await expect.poll(() => header.locator(".chat-pane__palette-open").count()).toBe(0);
+    await header.locator(".chat-header-session-menu__trigger").click();
+    await page.getByText("Open command palette", { exact: true }).click();
+    await page.locator(".cmd-palette__input").waitFor({ state: "visible" });
+  });
+
+  it("opens the mobile drawer by swipe across the full mobile-layout range", async () => {
+    const page = await openPage({ hasTouch: true, height: 393, width: 852 });
+    const shell = page.locator(".shell");
+    await expect.poll(() => shell.getAttribute("class")).toContain("shell--mobile-nav");
+
+    await page.locator(".content").evaluate((content) => {
+      const touch = (clientX: number, clientY: number) =>
+        new Touch({
+          identifier: 1,
+          target: content,
+          clientX,
+          clientY,
+          pageX: clientX,
+          pageY: clientY,
+          screenX: clientX,
+          screenY: clientY,
+        });
+      content.dispatchEvent(
+        new TouchEvent("touchstart", {
+          bubbles: true,
+          composed: true,
+          touches: [touch(24, 180)],
+          changedTouches: [touch(24, 180)],
+        }),
+      );
+      content.dispatchEvent(
+        new TouchEvent("touchmove", {
+          bubbles: true,
+          cancelable: true,
+          composed: true,
+          touches: [touch(210, 184)],
+          changedTouches: [touch(210, 184)],
+        }),
+      );
+      content.dispatchEvent(
+        new TouchEvent("touchend", {
+          bubbles: true,
+          composed: true,
+          touches: [],
+          changedTouches: [touch(210, 184)],
+        }),
+      );
+    });
+
+    await expect.poll(() => shell.getAttribute("class")).toContain("shell--nav-drawer-open");
+    await expect.poll(() => page.locator(".shell-nav.nav-drawer").isVisible()).toBe(true);
+  });
+});
