+++ ui/src/e2e/native-nav-sidebar-toggle.e2e.test.ts
@@ -620,116 +620,6 @@ suite.define(() => {
     }
   });
 
-  it("keeps only history controls in the Settings titlebar", async () => {
-    const page = await openPage({ webChrome: true });
-    const response = await page.goto(`${suite.server.baseUrl}settings/general`);
-    expect(response?.status()).toBe(200);
-
-    const toolbar = page.locator(".macos-titlebar-controls");
-    await expect.poll(() => toolbar.isVisible()).toBe(true);
-    await expect.poll(() => toolbar.getByRole("button").count()).toBe(2);
-    await expect.poll(() => toolbar.getByRole("button", { name: "Back" }).isVisible()).toBe(true);
-    await expect
-      .poll(() => toolbar.getByRole("button", { name: "Forward" }).isVisible())
-      .toBe(true);
-    await expect
-      .poll(() => toolbar.getByRole("button", { name: "Expand sidebar" }).count())
-      .toBe(0);
-    await expect
-      .poll(() => toolbar.getByRole("button", { name: "Open command palette" }).count())
-      .toBe(0);
-    await expect.poll(() => toolbar.getByRole("button", { name: "New session" }).count()).toBe(0);
-  });
-
-  it("keeps the document root scroll-locked in the Settings takeover", async () => {
-    const page = await openPage({ webChrome: true });
-    const response = await page.goto(`${suite.server.baseUrl}settings/general`);
-    expect(response?.status()).toBe(200);
-    await page.locator(".settings-sidebar").waitFor({ state: "visible" });
-
-    // WKWebView scrolls the document whenever it overflows, dragging the
-    // settings sidebar and content along. Force overflow the way stray
-    // content would, then confirm the root refuses to move.
-    const metrics = await page.evaluate(() => {
-      const spacer = document.createElement("div");
-      spacer.style.height = "3000px";
-      document.body.append(spacer);
-      window.scrollTo(0, 500);
-      document.documentElement.scrollTop = 500;
-      document.body.scrollTop = 500;
-      return {
-        bodyScrollTop: document.body.scrollTop,
-        htmlScrollTop: document.documentElement.scrollTop,
-        rootScrollY: window.scrollY,
-      };
-    });
-    expect(metrics).toEqual({ bodyScrollTop: 0, htmlScrollTop: 0, rootScrollY: 0 });
-  });
-
-  it("keeps drawer and search reachable from the narrow chat title bar", async () => {
-    const page = await openPage({ nativeNav: false, width: 900 });
-    const header = page.locator(".chat-pane__header").first();
-    await expect
-      .poll(() => page.locator(".shell").getAttribute("class"))
-      .toContain("shell--merged-chat-chrome");
-    await expect.poll(() => page.locator(".topbar").isVisible()).toBe(false);
-    await expect
-      .poll(() => header.getByRole("button", { name: "Expand sidebar" }).isVisible())
-      .toBe(true);
-    await expect.poll(() => header.locator(".chat-pane__palette-open").count()).toBe(0);
-    await header.locator(".chat-header-session-menu__trigger").click();
-    await page.getByText("Open command palette", { exact: true }).click();
-    await page.locator(".cmd-palette__input").waitFor({ state: "visible" });
-  });
-
-  it("opens the mobile drawer by swipe across the full mobile-layout range", async () => {
-    const page = await openPage({ hasTouch: true, height: 393, nativeNav: false, width: 852 });
-    const shell = page.locator(".shell");
-    await expect.poll(() => shell.getAttribute("class")).toContain("shell--mobile-nav");
-
-    await page.locator(".content").evaluate((content) => {
-      const touch = (clientX: number, clientY: number) =>
-        new Touch({
-          identifier: 1,
-          target: content,
-          clientX,
-          clientY,
-          pageX: clientX,
-          pageY: clientY,
-          screenX: clientX,
-          screenY: clientY,
-        });
-      content.dispatchEvent(
-        new TouchEvent("touchstart", {
-          bubbles: true,
-          composed: true,
-          touches: [touch(24, 180)],
-          changedTouches: [touch(24, 180)],
-        }),
-      );
-      content.dispatchEvent(
-        new TouchEvent("touchmove", {
-          bubbles: true,
-          cancelable: true,
-          composed: true,
-          touches: [touch(210, 184)],
-          changedTouches: [touch(210, 184)],
-        }),
-      );
-      content.dispatchEvent(
-        new TouchEvent("touchend", {
-          bubbles: true,
-          composed: true,
-          touches: [],
-          changedTouches: [touch(210, 184)],
-        }),
-      );
-    });
-
-    await expect.poll(() => shell.getAttribute("class")).toContain("shell--nav-drawer-open");
-    await expect.poll(() => page.locator(".shell-nav.nav-drawer").isVisible()).toBe(true);
-  });
-
   it("keeps overlay motion anchored to its owning interaction", async () => {
     const page = await openPage({ nativeNav: false });
